sys-auth/elogind : Add config and init files.
With elogind relying only on dbus activation the following
problem occurs with SDDM:
The sddm-greeter uses pam_elogind.so for session management
which causes dbus to start elogind as it is not yet running
when xdm starts sddm.
SDDM itself, however, got no answer about its possibilities
and fails to get the proper actions for shutdown, reboot,
suspend and hibernate. The sddm-greeter then only shows
buttons for shutdown and reboot, which do not work.
Plasma fails to shutdown or reboot the machine, too, as the
process in doing so stops when sddm takes over after the
user has logged out from Plasma.

Solution:

New init and conf files allow to add elogind to the default
runlevel. It needs dbus and is started before xdm.
With this procedure elogind is up and running when sddm is
started, and shows all available buttons. These now work as
expected.

Note:

I am not sure it is still neccessary to patch
/usr/share/dbus-1/system-services/org.freedesktop.login1.service
to have an explicit entry in the "Exec" line, as dbus no
longer needs to start elogind when it is already running.