app-arch/lhasa: Added V 0.3.1 for milkytracker (desc below)
Free Software LHA implementation
Lhasa is a Free Software replacement for the Unix LHA tool, for
decompressing .lzh (LHA / LHarc) and .lzs (LArc) archives. The
backend for the tool is a library, so that it can be reused for
other purposes.

Lhasa aims to be compatible with as many types of lzh/lzs archives as
possible. It also aims to generate the same output as the (non-free)
Unix LHA tool, so that it will act as a drop-in free replacement.

License
ISC license (functionally equivalent to the BSD / MIT licenses).

Authors
Simon Howard (fraggle@removethistoemailme.gmail.com)

Package-Manager: Portage-3.0.30, Repoman-3.0.3
Signed-off-by: Sven Eden <sven.eden@prydeworx.com>